data-processing/requirements.txt
# RabbitMQ
pika==1.3.2

# API içi
fastapi==0.115.12
uvicorn==0.34.0


# veri isleme
pydantic==2.11.3
python-geohash==0.8.5
geopy==2.4.1

# veri tabanı
influxdb==5.3.2

#logging
loguru==0.7.3

# cevresel degiskenler
python-dotenv==1.1.0

#test 
pytest==8.3.5

#influxdb
influxdb-client==1.48.0